added get surcharge method to insurance calculator

+++ CoolBlue.Products.Application/Common/Interfaces/IInsuranceService.cs
@@ -4,8 +4,8 @@ namespace CoolBlue.Products.Application.Common.Interfaces
 {
     public interface IInsuranceService
     {
-        Task<float> CalculateInsurance(int productId, CancellationToken cancellationToken);
-        Task<float> CalculateBatchInsurance(List<int> productIdList, CancellationToken cancellationToken);
+        Task<double> CalculateInsurance(int productId, CancellationToken cancellationToken);
+        Task<double> CalculateBatchInsurance(List<int> productIdList, CancellationToken cancellationToken);
         Task AddSurcharge(int productTypeId, double surchargeRate, CancellationToken cancellationToken);
     }
 }
